MODIFICACION INTERMEDIA PROYECTO

+++ MS-pagos/src/main/resources/db/migration/V1__crear_tabla_pagos.sql
@@ -2,7 +2,7 @@ CREATE TABLE IF NOT EXISTS pagos (
     id                 BIGINT         AUTO_INCREMENT PRIMARY KEY,
     pedido_id          BIGINT         NOT NULL COMMENT 'FK logica a MS-pedidos',
     transaccion_id     VARCHAR(100)   UNIQUE COMMENT 'UUID generado en procesarPago',
-    monto              DECIMAL(10, 2) NOT NULL,
+    monto              FLOAT(10, 2) NOT NULL,
     moneda             VARCHAR(10)    NOT NULL DEFAULT 'CLP',
     metodo             VARCHAR(50)    NOT NULL COMMENT 'EFECTIVO | DEBITO | CREDITO | TRANSFERENCIA',
     estado             VARCHAR(30)    NOT NULL COMMENT 'PROCESANDO | COMPLETADO | RECHAZADO',
